Created appliction and added servicies for sending requests on the server
