--- app/app/application.ini
@@ -1,9 +1,0 @@
-[options]
-primary_palette = Amber
-accent_palette = Amber
-theme_style = Dark
-primary_hue = 500
-
-[global]
-iterable_count = 5
-
