+++ app/exceptions/__init__.py
@@ -1,0 +1,2 @@
+class NoSecretKeyError(Exception):
+    ''' Секретный ключ не обнаружен '''
