+++ app/kv/screens/app_screen.kv
@@ -2,7 +2,7 @@
 #:import LOGO_IMG config.LOGO_IMG
 
 
-<FDUIManager>:
+<FDNavigation>:
 	ScreenManager:
 		id: screen_manager
