+++ app/requirements.txt
@@ -1,0 +1,16 @@
+certifi==2025.11.12
+charset-normalizer==3.4.4
+docutils==0.20.1
+idna==3.11
+Kivy==2.1.0
+kivy-deps.angle==0.3.3
+kivy-deps.glew==0.3.1
+kivy-deps.sdl2==0.4.5
+Kivy-Garden==0.1.5
+kivymd==0.104.2
+pillow==10.4.0
+Pygments==2.19.2
+pypiwin32==223
+pywin32==311
+requests==2.32.4
+urllib3==2.2.3
