+++ app/ui/screens/calls.py
@@ -9,7 +9,7 @@ from kivymd.uix.dialog import MDDialog
 from kivymd.uix.boxlayout import MDBoxLayout
 from kivymd.uix.button import MDRaisedButton, MDFlatButton
 
-from . import BaseScreen
+from .base import BaseScreen
 from data_base.model import Human, Short, Emergency, Calls, Rank, Worktype
 from data_base.manager import write_entry, get_by_id
 from exceptions.data_base import DBAddError, DBCommitError
