+++ app/ui/screens/history.py
@@ -4,7 +4,7 @@ from kivy.metrics import dp
 from kivymd.uix.dialog import MDDialog
 from kivymd.uix.button import MDRaisedButton
 
-from . import BaseScrollScreen
+from .base import BaseScrollScreen
 from data_base.model import Calls, Emergency
 from app.path_manager import PathManager
 from ui.layout.label_layout import FDLabelLayout
